Refactor: Move handleCellDoubleClick logic to SelectableCellMixin
